+++ packages/api/e2e/amino.spec.ts
@@ -7,6 +7,8 @@ import {
   MsgCreateProject,
   MsgRetire,
   MsgSend,
+  MsgBridge,
+  MsgMintBatchCredits,
 } from '../src/generated/regen/ecocredit/v1/tx';
 import { StdFee } from '@cosmjs/amino/build/signdoc';
 import { Secp256k1HdWallet } from '@cosmjs/amino/build/secp256k1hdwallet';
@@ -94,7 +96,7 @@ xdescribe('RegenApi with tendermint connection - Amino Tests', () => {
 
       const TEST_MSG_SEND = MsgSend.fromPartial({
         sender: TEST_ADDRESS,
-        recipient: TEST_ADDRESS,
+        recipient: TEST_BUYER_ADDRESS,
         credits: [
           {
             batchDenom: TEST_BATCH_DENOM,
@@ -110,12 +112,13 @@ xdescribe('RegenApi with tendermint connection - Amino Tests', () => {
 
       const TEST_MSG_SEND = MsgSend.fromPartial({
         sender: TEST_ADDRESS,
-        recipient: TEST_ADDRESS,
+        recipient: TEST_BUYER_ADDRESS,
         credits: [
           {
             batchDenom: TEST_BATCH_DENOM,
             retiredAmount: MIN_CREDIT_AMOUNT,
             retirementJurisdiction: 'US-WA',
+            retirementReason: 'test',
           },
         ],
       });
@@ -127,13 +130,14 @@ xdescribe('RegenApi with tendermint connection - Amino Tests', () => {
 
       const TEST_MSG_SEND = MsgSend.fromPartial({
         sender: TEST_ADDRESS,
-        recipient: TEST_ADDRESS,
+        recipient: TEST_BUYER_ADDRESS,
         credits: [
           {
             batchDenom: TEST_BATCH_DENOM,
             tradableAmount: MIN_CREDIT_AMOUNT,
             retiredAmount: MIN_CREDIT_AMOUNT,
             retirementJurisdiction: 'US-WA',
+            retirementReason: 'test',
           },
         ],
       });
@@ -193,6 +197,7 @@ xdescribe('RegenApi with tendermint connection - Amino Tests', () => {
             tradableAmount: '1.503',
             retiredAmount: '1.503',
             retirementJurisdiction: 'US-OR',
+            retirementReason: 'test',
           },
         ],
         startDate: startDate,
@@ -232,7 +237,71 @@ xdescribe('RegenApi with tendermint connection - Amino Tests', () => {
 
       await runAminoTest(msgClient, TEST_MSG_CREATE_BATCH);
     });
+    it('should sign and broadcast MsgMintBatch using legacy amino sign mode', async () => {
+      const { msgClient } = await connect();
+
+      const TEST_MSG_MINT_BATCH = MsgMintBatchCredits.fromPartial({
+        issuer: TEST_ADDRESS,
+        batchDenom: TEST_BATCH_DENOM,
+        issuance: [
+          {
+            recipient: TEST_ADDRESS,
+            tradableAmount: '1.503',
+            retiredAmount: '1.503',
+            retirementJurisdiction: 'US',
+            retirementReason: 'test',
+          },
+        ],
+        originTx: {
+          id: makeEthTxHash(),
+          source: 'polygon',
+          contract: makeEthContract(),
+          note: 'bridged from another chain',
+        },
+      });
+
+      await runAminoTest(msgClient, TEST_MSG_MINT_BATCH);
+    });
+    it('should sign and broadcast MsgMintBatch with default values using legacy amino sign mode', async () => {
+      const { msgClient } = await connect();
+
+      const TEST_MSG_MINT_BATCH = MsgMintBatchCredits.fromPartial({
+        issuer: TEST_ADDRESS,
+        batchDenom: TEST_BATCH_DENOM,
+        issuance: [
+          {
+            recipient: TEST_ADDRESS,
+            tradableAmount: '1.503',
+          },
+        ],
+        originTx: {
+          id: makeEthTxHash(),
+          source: 'polygon',
+          contract: makeEthContract(),
+          note: 'bridged from another chain',
+        },
+      });
+
+      await runAminoTest(msgClient, TEST_MSG_MINT_BATCH);
+    });
     it('should sign and broadcast MsgRetire using legacy amino sign mode', async () => {
+      const { msgClient } = await connect();
+
+      const TEST_MSG_RETIRE = MsgRetire.fromPartial({
+        owner: TEST_ADDRESS,
+        credits: [
+          {
+            batchDenom: TEST_BATCH_DENOM,
+            amount: '0.000001',
+          },
+        ],
+        jurisdiction: 'US-OR',
+        reason: 'test',
+      });
+
+      await runAminoTest(msgClient, TEST_MSG_RETIRE);
+    });
+    it('should sign and broadcast MsgRetire with default values using legacy amino sign mode', async () => {
       const { msgClient } = await connect();
 
       const TEST_MSG_RETIRE = MsgRetire.fromPartial({
@@ -263,6 +332,23 @@ xdescribe('RegenApi with tendermint connection - Amino Tests', () => {
       });
 
       await runAminoTest(msgClient, TEST_MSG_CANCEL);
+    });
+    it('should sign and broadcast MsgBridge using legacy amino sign mode', async () => {
+      const { msgClient } = await connect();
+
+      const TEST_MSG_BRIDGE = MsgBridge.fromPartial({
+        owner: TEST_ADDRESS,
+        recipient: makeEthContract(),
+        target: 'polygon',
+        credits: [
+          {
+            batchDenom: TEST_BATCH_DENOM,
+            amount: MIN_CREDIT_AMOUNT,
+          },
+        ],
+      });
+
+      await runAminoTest(msgClient, TEST_MSG_BRIDGE);
     });
   });
   describe('Signing and broadcasting Basket txs using legacy amino sign mode', () => {
@@ -305,7 +391,20 @@ xdescribe('RegenApi with tendermint connection - Amino Tests', () => {
       const TEST_BASKET_MSG_TAKE = MsgTake.fromPartial({
         owner: TEST_ADDRESS,
         basketDenom: 'eco.uC.NCT',
-        amount: '1000000',
+        amount: '100000',
+        retirementJurisdiction: 'US-CO 98765',
+        retireOnTake: true,
+        retirementReason: 'test',
+      });
+
+      await runAminoTest(msgClient, TEST_BASKET_MSG_TAKE);
+    });
+    it('should sign and broadcast MsgTake with default values', async () => {
+      const { msgClient } = await connect();
+      const TEST_BASKET_MSG_TAKE = MsgTake.fromPartial({
+        owner: TEST_ADDRESS,
+        basketDenom: 'eco.uC.NCT',
+        amount: '100000',
         retirementJurisdiction: 'US-CO 98765',
         retireOnTake: true,
       });
@@ -392,6 +491,7 @@ xdescribe('RegenApi with tendermint connection - Amino Tests', () => {
             bidPrice: { denom: 'uregen', amount: '1000000' },
             disableAutoRetire: false,
             retirementJurisdiction: 'US',
+            retirementReason: 'test',
           },
         ],
       });