+++ packages/api/src/tx/modules/ecocredit/v1/mint_batch_credits_amino.ts
@@ -7,7 +7,7 @@ import {
 } from '../../../../generated/regen/ecocredit/v1/types';
 import { AminoBatchIssuance, AminoOriginTx } from './create_batch_amino';
 
-const msgMintBatchCreditsAminoType = 'regen.core/MsgMintBatchCredits';
+const msgMintBatchCreditsAminoType = 'regen/MsgMintBatchCredits';
 
 export const mintBatchCreditsTypeUrl = '/' + MsgMintBatchCredits.$type;
 
@@ -41,10 +41,11 @@ export function mintBatchCreditsConverter(): AminoConverter {
         batch_denom: batchDenom,
         issuance: issuance.map(i => {
           return {
-            recipient: i.recipient,
-            tradable_amount: i.tradableAmount,
-            retired_amount: i.retiredAmount,
-            retirement_jurisdiction: i.retirementJurisdiction,
+            recipient: i.recipient || undefined,
+            tradable_amount: i.tradableAmount || undefined,
+            retired_amount: i.retiredAmount || undefined,
+            retirement_jurisdiction: i.retirementJurisdiction || undefined,
+            retirement_reason: i.retirementReason || undefined,
           };
         }),
         origin_tx: originTx && {
@@ -69,6 +70,7 @@ export function mintBatchCreditsConverter(): AminoConverter {
             tradableAmount: i.tradable_amount || '',
             retiredAmount: i.retired_amount || '',
             retirementJurisdiction: i.retirement_jurisdiction || '',
+            retirementReason: i.retirement_reason || '',
           };
         }),
         originTx: origin_tx && {