+++ CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/regen-network/regen-js/compare/v0.3.0...v0.6.0) (2023-03-27)
+
+
+### Features
+
+* Add amino signing support to regen-ledger msgs ([#38](https://github.com/regen-network/regen-js/issues/38)) ([84d6b5a](https://github.com/regen-network/regen-js/commit/84d6b5a7b344f893d80ba54b99067677cfdb36ef)), closes [#39](https://github.com/regen-network/regen-js/issues/39) [#43](https://github.com/regen-network/regen-js/issues/43) [#44](https://github.com/regen-network/regen-js/issues/44) [#45](https://github.com/regen-network/regen-js/issues/45) [#46](https://github.com/regen-network/regen-js/issues/46) [#48](https://github.com/regen-network/regen-js/issues/48) [#50](https://github.com/regen-network/regen-js/issues/50) [#49](https://github.com/regen-network/regen-js/issues/49) [#47](https://github.com/regen-network/regen-js/issues/47) [#53](https://github.com/regen-network/regen-js/issues/53) [#52](https://github.com/regen-network/regen-js/issues/52) [#54](https://github.com/regen-network/regen-js/issues/54) [#55](https://github.com/regen-network/regen-js/issues/55) [#56](https://github.com/regen-network/regen-js/issues/56)
+* add bridge_amino and test ([#61](https://github.com/regen-network/regen-js/issues/61)) ([cb0d339](https://github.com/regen-network/regen-js/commit/cb0d33904c12b61a06def9aa85a56c96a633743e))
+* add transfer ibc proto ([#40](https://github.com/regen-network/regen-js/issues/40)) ([61c6cd2](https://github.com/regen-network/regen-js/commit/61c6cd2d4b3903da8bbb5d771dd815bdeb1c1000))
+* update ecocredit amino types ([#64](https://github.com/regen-network/regen-js/issues/64)) ([6c8fabb](https://github.com/regen-network/regen-js/commit/6c8fabb8c24c977446f879b861c8e338d4b58b60))
+* update proto files for regen-ledger v5.0 and cosmos-sdk v0.46.7 ([#66](https://github.com/regen-network/regen-js/issues/66)) ([5feaeed](https://github.com/regen-network/regen-js/commit/5feaeeda4f32eb3379f0b5c1b2dc7557c2e08b46))
+
+
+
+
+
+# [0.5.0](https://github.com/regen-network/regen-js/compare/v0.3.0...v0.5.0) (2023-01-16)
+
+
+### Features
+
+* Add amino signing support to regen-ledger msgs ([#38](https://github.com/regen-network/regen-js/issues/38)) ([84d6b5a](https://github.com/regen-network/regen-js/commit/84d6b5a7b344f893d80ba54b99067677cfdb36ef)), closes [#39](https://github.com/regen-network/regen-js/issues/39) [#43](https://github.com/regen-network/regen-js/issues/43) [#44](https://github.com/regen-network/regen-js/issues/44) [#45](https://github.com/regen-network/regen-js/issues/45) [#46](https://github.com/regen-network/regen-js/issues/46) [#48](https://github.com/regen-network/regen-js/issues/48) [#50](https://github.com/regen-network/regen-js/issues/50) [#49](https://github.com/regen-network/regen-js/issues/49) [#47](https://github.com/regen-network/regen-js/issues/47) [#53](https://github.com/regen-network/regen-js/issues/53) [#52](https://github.com/regen-network/regen-js/issues/52) [#54](https://github.com/regen-network/regen-js/issues/54) [#55](https://github.com/regen-network/regen-js/issues/55) [#56](https://github.com/regen-network/regen-js/issues/56)
+* add bridge_amino and test ([#61](https://github.com/regen-network/regen-js/issues/61)) ([cb0d339](https://github.com/regen-network/regen-js/commit/cb0d33904c12b61a06def9aa85a56c96a633743e))
+* add transfer ibc proto ([#40](https://github.com/regen-network/regen-js/issues/40)) ([61c6cd2](https://github.com/regen-network/regen-js/commit/61c6cd2d4b3903da8bbb5d771dd815bdeb1c1000))
+* update ecocredit amino types ([#64](https://github.com/regen-network/regen-js/issues/64)) ([6c8fabb](https://github.com/regen-network/regen-js/commit/6c8fabb8c24c977446f879b861c8e338d4b58b60))
+
+
+
+
+
 # [0.4.0](https://github.com/regen-network/regen-js/compare/v0.4.0-alpha1.5...v0.4.0) (2022-10-05)
 
 **Note:** Version bump only for package regen-js